chore: move ProductsOutput to dto folder
