+++ src/domain/entities/product/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './product-output'
+export * from './products-output'
