--- src/domain/usecases/product/common/index.ts
@@ -1,1 +1,0 @@
-export * from './products-output'
